+++ valkey_glide_cluster.c
@@ -1121,6 +1121,10 @@ TIME_METHOD_IMPL(ValkeyGlideCluster)
 LOLWUT_METHOD_IMPL(ValkeyGlideCluster)
 /* }}} */
 
+/* {{{ proto int|array ValkeyGlideCluster::lastSave([mixed $route]) */
+LASTSAVE_METHOD_IMPL(ValkeyGlideCluster)
+/* }}} */
+
 /* {{{ proto string ValkeyGlideCluster::randomkey(string key)
  *     proto string ValkeyGlideCluster::randomkey(array host_port) */
 RANDOMKEY_METHOD_IMPL(ValkeyGlideCluster)