+++ valkey_glide_commands_common.h
@@ -250,6 +250,7 @@ int execute_reset_command(zval* object, int argc, zval* return_value, zend_class
 int execute_save_command(zval* object, int argc, zval* return_value, zend_class_entry* ce);
 int execute_time_command(zval* object, int argc, zval* return_value, zend_class_entry* ce);
 int execute_lolwut_command(zval* object, int argc, zval* return_value, zend_class_entry* ce);
+int execute_lastsave_command(zval* object, int argc, zval* return_value, zend_class_entry* ce);
 int execute_scan_command(zval* object, int argc, zval* return_value, zend_class_entry* ce);
 int execute_cluster_scan_command(const void* glide_client,
                                  char**      cursor,
@@ -579,6 +580,9 @@ int execute_unlink_command(zval* object, int argc, zval* return_value, zend_clas
 
 #define LOLWUT_METHOD_IMPL(class_name) \
     STANDARD_METHOD_IMPL(class_name, lolwut, execute_lolwut_command)
+
+#define LASTSAVE_METHOD_IMPL(class_name) \
+    STANDARD_METHOD_IMPL(class_name, lastSave, execute_lastsave_command)
 
 #define SSCAN_METHOD_IMPL(class_name) STANDARD_METHOD_IMPL(class_name, sscan, execute_sscan_command)